refactor(link): expose resolver context only
